requirements.txt
# ── API server ──────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
aiofiles==23.2.1

# ── LangChain ───────────────────────────────────────────────────────────────
langchain==0.2.6
langchain-groq==0.1.9

# ── PDF extraction ──────────────────────────────────────────────────────────
pdfplumber==0.11.1
pymupdf==1.24.5

# ── Embeddings & vector search ──────────────────────────────────────────────
sentence-transformers==3.0.1
faiss-cpu==1.8.0
rank-bm25==0.2.2

# ── ML utilities ────────────────────────────────────────────────────────────
numpy==1.26.4
nltk==3.8.1
torch==2.3.0

# ── Streamlit UI ────────────────────────────────────────────────────────────
streamlit==1.35.0
requests==2.32.3

# ── Config & logging ────────────────────────────────────────────────────────
python-dotenv==1.0.1
pydantic==2.7.4
loguru==0.7.2
